Finish the graph
Break it into a few templates to make it easier to theme
